REGRESSION (r201629): Weird button glitching on github.com https://bugs.webkit.org/show_bug.cgi?id=159031 rdar://problem/26880332
Reviewed by Tim Horton.

Source/WebCore:

r201629 changed the logic slightly when creating an image buffer for a scaled context;
it set the buffer context's scale to the scale in the source context, but this failed
to take into account the rounding up of the buffer size, which the old code did.

Fix by reverting to the old behavior.

Since buffer sizes can only be integral, changed compatibleBufferSize() to return
an IntSize.

Test: fast/backgrounds/scaled-gradient-background.html

* platform/graphics/ImageBuffer.cpp:
(WebCore::ImageBuffer::createCompatibleBuffer):
(WebCore::ImageBuffer::compatibleBufferSize):
* platform/graphics/ImageBuffer.h:
* platform/graphics/IntRect.h:
(WebCore::IntRect::area):
* platform/graphics/IntSize.h:
(WebCore::IntSize::area): Make this return an unsigned.

LayoutTests:

* fast/backgrounds/scaled-gradient-background-expected.html: Added.
* fast/backgrounds/scaled-gradient-background.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@202348 268f45cc-cd09-0410-ab3c-d52691b4dbfc